// src/types/aircraft.ts
// Shared Aircraft type definition
export interface Aircraft {
  name: string;
  user: string;
  startYear: number;
  endYear: number;
  totalCount: number;
  klu: number;
  mld: number;
  mlknil: number;
  notes: string;
  museum: string;
  wreckAssessment: number;
  foto: string;
  localImage: string;
  preservationStatus: string;
  aircraftType: string;
}